+++ clang/compilecommand_gen.go
@@ -18,8 +18,8 @@ func (cc CompileCommand) Directory() string {
 }
 
 // Get the number of arguments in the compiler invocation.
-func (cc CompileCommand) NumArgs() uint16 {
-	return uint16(C.clang_CompileCommand_getNumArgs(cc.c))
+func (cc CompileCommand) NumArgs() uint32 {
+	return uint32(C.clang_CompileCommand_getNumArgs(cc.c))
 }
 
 /*
@@ -28,7 +28,7 @@ func (cc CompileCommand) NumArgs() uint16 {
 	Invariant :
 	- argument 0 is the compiler executable
 */
-func (cc CompileCommand) Arg(i uint16) string {
+func (cc CompileCommand) Arg(i uint32) string {
 	o := cxstring{C.clang_CompileCommand_getArg(cc.c, C.uint(i))}
 	defer o.Dispose()
