+++ clang/turesourceusageentry_gen.go
@@ -12,6 +12,6 @@ func (turue TUResourceUsageEntry) Kind() TUResourceUsageKind {
 	return TUResourceUsageKind(turue.c.kind)
 }
 
-func (turue TUResourceUsageEntry) Amount() uint32 {
-	return uint32(turue.c.amount)
+func (turue TUResourceUsageEntry) Amount() uint64 {
+	return uint64(turue.c.amount)
 }
